% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{breakpoints}
\alias{breakpoints}
\title{Fama/French variable breakpoints}
\format{A \linkS4class{data.table} with variables:
\itemize{
  \item{variable: Fama/French variable name. See the "\link{variables}" dataset in this
  package as well as \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/variable_definitions.html}
  for details.}
  \item{frequency: variable observation frequency.}
  \item{percentile: percentile breakpoint. "# negative" and "# positive" count the number of negative and
  positive occurences for a given period repectively.}
  \item{period: observation period.}
  \item{value: corresponding variable breakpoint value.}
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.
}
\usage{
breakpoints
}
\description{
Variable breakpoints time series. The variable breakpoints are
  used to construct the Fama/French portfolios and factors.
}
\keyword{datasets}
