% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFresearch-package.R
\docType{package}
\name{FFresearch-package}
\alias{FFresearch}
\alias{FFresearch-package}
\title{FFresearch: Eugene Fama & Kenneth French research data}
\description{
Fama/French asset pricing research data conveniently packaged for consumption in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bautheac.github.io/FFresearch}
  \item Report bugs at \url{https://github.com/bautheac/FFresearch/issues}
}

}
\author{
\strong{Maintainer}: Olivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
