% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{portfolios_trivariate}
\alias{portfolios_trivariate}
\title{Fama/French three-way sorts}
\format{
A \linkS4class{data.table} with variables:
\itemize{
  \item{region: geographical market considered for factor construction.}
  \item{frequency: field observation frequency.}
  \item{sort variable 1: first sort variable used for portfolio construction.}
  \item{sort variable 2: second sort variable used for portfolio construction.}
  \item{sort variable 3: third sort variable used for portfolio construction.}
  \item{diviend: Y (included) or N (excluded).}
  \item{weights: portoflio weighting method.}
  \item{portfolio: corresponding portoflio.}
  \item{field: variable observed for given portfolio.}
  \item{period: observation period.}
  \item{value: corresponding field value.}
}
}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.
}
\usage{
portfolios_trivariate
}
\description{
Return time series for Fama/French portfolios formed on
  sorting on three variables.
}
\keyword{datasets}
