% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{variables}
\alias{variables}
\title{Fama/French variables}
\format{
A \link[tibble]{tibble} with variables:
\itemize{
  \item{name: variable name.}
  \item{symbol: symbol used to refer to a given variable in the various datasets
  contained in the package.}
  \item{description: variable descriptive elements including contruction methods.
  See\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/variable_definitions.html}
  for more details.}
}
}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.
}
\usage{
variables
}
\description{
Helper dataset. Provides details on Fama/French variables,
  including contruction methods.
}
\keyword{datasets}
